move get_full_redirect_app_url to server
