+++ packages/kbn-reporting/server/build_kibana_path.ts
