+++ packages/kbn-reporting/server/get_full_redirect_app_url.test.ts
