+++ packages/kbn-reporting/server/index.ts
@@ -6,18 +6,19 @@
  * Side Public License, v 1.
  */
 
-export { getFieldFormats, setFieldFormats } from './services';
-export { decryptJobHeaders } from './decrypt_job_headers';
-export { getCustomLogo } from './get_custom_logo';
-export { generatePdfObservable } from './generate_pdf';
-export { validateUrls } from './validate_urls';
-export { getFullUrls } from './get_full_urls';
-export { getAbsoluteUrlFactory } from './get_absolute_url';
-export { generatePngObservable } from './generate_png';
-export { generatePdfObservableV2 } from './generate_pdf_v2';
-export { buildKibanaPath } from '@kbn/reporting-common/build_kibana_path';
+export { buildKibanaPath } from './build_kibana_path';
 export { cryptoFactory } from './crypto';
+export { decryptJobHeaders } from './decrypt_job_headers';
 export { ExportType } from './export_type';
 export type { BaseExportTypeSetupDeps, BaseExportTypeStartDeps } from './export_type';
+export { generatePdfObservable } from './generate_pdf';
+export { generatePdfObservableV2 } from './generate_pdf_v2';
+export { generatePngObservable } from './generate_png';
+export { getAbsoluteUrlFactory } from './get_absolute_url';
+export { getCustomLogo } from './get_custom_logo';
+export { getFullRedirectAppUrl } from './get_full_redirect_app_url';
+export { getFullUrls } from './get_full_urls';
 export { getTracker } from './pdf_tracker';
+export { getFieldFormats, setFieldFormats } from './services';
 export * from './types';
+export { validateUrls } from './validate_urls';
